Memoro provides memory to dlsym for error messages
On glibc 2.23+, the function dlsym() allocates memory to store its
error message, replacing the older behavior of using a statically
allocated memory that could not find large error messages.
https://sourceware.org/bugzilla/show_bug.cgi?id=19509

On Linux the functions __isoc99_*() are instrumented, but they might
not be linked to the binary. This leads to some error from dlsym() when
it tries to load the real functions behind the interceptors.
https://bugs.llvm.org/show_bug.cgi?id=27310

Those error issued a call to malloc() while Memoro was still in its init
phase and failed the ENSURE_MEMORO_INITIED() check.
Now Memoro offer a statically allocated buffer for allocations before it
has been fully initialized and doesn't fail the check anymore.
https://reviews.llvm.org/D20235