Disable iOS target for Memoro
This is a proper fix for a87b20e issue.

We need thread local variables for Memoro's run-time.
It is used for performance reason in the sampling process and
is for storing the stack's stop in the stack filter.

Thread local variables are supported in recent version of macOS and
in bleeding edge version of iOS. Previous versions don't support
TLS and prevent the compilation of Memoro's run-time.

MacOS supports the feature since long enough, but it's not the case of
iOS. This commit disable the Memoro's run-time targeting iOS, to prevent
compilation issue with this target.
Memoro on iOS is such an edge case that we don't have to support it yet.